refactor, moved getNextTaskLsid to LSIDManager class GP-6212, GP-6220
